moved navbar components into its separate folder
