+++ client/src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Navbar/Navbar";
 import "./App.css";
 
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
