+++ client/src/pages/Consulting.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Consulting() {
+  return (
+    <>
+      <h1 className="">timetables</h1>
+    </>
+  );
+}
