+++ client/src/pages/Marketing.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Marketing() {
+  return (
+    <>
+      <h1 className='marketing'>MARKETING</h1>
+    </>
+  );
+}
