+++ client/src/pages/Products.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import '../../App.css';
+
+export default function Products() {
+  return <h1 className='products'>PRODUCTS</h1>;
+}
