+++ client/src/pages/Services.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "../../App.css";
+
+export default function Services() {
+  return <h1 className=""></h1>;
+}
