+++ client/src/pages/WhatIsAutism.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import '../../App.css';
+
+export default function WhatIsAutism() {
+  return <h1 className='sign-up'>LIKE & SUBSCRIBE</h1>;
+}
